was/requirements.txt
Django==1.10
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
Pillow==5.0.0
PyJWT==1.5.3
pytz==2017.3
